Expliciated pipeline stages
Created a primitive->tansform->clip->draw stages for pipeline. Instead
of directly drawing primitives we now queue them with
QueueTransformClipX function. X for given primitive. During the queue
the lines are transformed and clipped by the set matrices and regions.
Also added the said matrices and regions, which are set with
SetProjectionTransform, SetCameraTransform and SetClipRectangle methods.