Added SpriteLoader and rewrote the eng. interface
Added a way to load animations via a single sprite image through
SpriteLoader. Changed the way components interact with each other.
Instead of having a local pointer to the Software Rasterizer, we now
have a single extern SoftwareRasterizer. Changed the responsibilities of
WindowsMain file. It's now only handling window creation. Initialization
of the application and software rasterizer is now delegated to the
GameEngine file.